==> notebook/02-data split.ipynb <==
import pandas
import numpy as np

data = pandas.read_csv('../data/comp_touzi_comp.csv')

df = data.set_index(data['date'])
df = df.sort_index()

df

data.src_ind.max(), data.dst_ind.max()

train_df, valid_df, test_df = df[:'2018-12-31'][['src_ind', 'dst_ind', 'date']], df['2019-01-01':'2019-12-31'][['src_ind', 'dst_ind', 'date']], df['2020-01-01':][['src_ind', 'dst_ind', 'date']]

len(train_df)/78114, len(valid_df)/78114, len(test_df)/78114

def negative_sampling(test_df, num_nodes, k=99):
    neg_src = np.repeat(test_df['src_ind'], k)
    neg_dst = np.random.randint(0, num_nodes, len(neg_src))
    test_negative_df = pandas.DataFrame({'src_ind': neg_src, 'dst_ind': neg_dst})
    return test_negative_df

test_negative_df = negative_sampling(test_df, data.dst_ind.max(), 99)

df1 = test_negative_df.set_index(['src_ind', 'dst_ind'])
df2 = df.set_index(['src_ind', 'dst_ind'])
df1.join(df2, how='inner')

len(df1), len(test_df)

train_df.to_csv('../tyc_cm/train_df.csv', index=False)
valid_df.to_csv('../tyc_cm/valid_df.csv', index=False)
test_df.to_csv('../tyc_cm/test_df.csv', index=False)
test_negative_df.to_csv('../tyc_cm/test_neg_df.csv', index=False)
