==> notebook/02-data-split.ipynb <==
import pandas as pd
import numpy as np
import random

path = '../data/tyc/'
data = pd.read_csv(path + 'comp_touzi_comp.csv')

df = data.set_index(data['date'])
df = df.sort_index()

df

print(f'user.max: {data.src_ind.max()}, item.max: {data.dst_ind.max()}')

src_set, dst_set = set(data.src_ind), set(data.dst_ind)
print(f'#user: {len(src_set)}, #item: {len(dst_set)}, user & item: {len(src_set&dst_set)}, user | item: {len(src_set|dst_set)}')

# print #records after date
def print_stat(dt):
    comps = set(df[dt:]['src_ind'])
    print(f'{dt}~: {len(comps)} comps, {len(df[dt:])} records')

print_stat('2020-01-01')
print_stat('2019-11-30')
print_stat('2019-06-30')
print_stat('2019-01-01')
print_stat('2018-01-01')

# remove names
df_ = df[['src_ind', 'dst_ind', 'date']]
df_.insert(3, 'label', 1)

train_df = df_[:'2018-12-31']
valid_df = df_['2019-01-01':'2019-12-31']
test_df = df_['2020-01-01':]

train_df

def negative_sampling(df, dst_set, k=99):
    neg_src = np.repeat(df['src_ind'], k)
    dst = list(dst_set)
    neg_dst = [random.choice(dst) for _ in range(len(neg_src))]
    test_neg_df = pd.DataFrame({'src_ind': neg_src, 'dst_ind': neg_dst, 'label': 0})
    return test_neg_df

def negative_sampling_by_user(df, dst_set, k=99):
    neg_src = np.repeat(list(set(df['src_ind'])), k)
    dst = list(dst_set)
    neg_dst = [random.choice(dst) for _ in range(len(neg_src))]
    test_neg_df = pd.DataFrame({'src_ind': neg_src, 'dst_ind': neg_dst, 'label': 0})
    return test_neg_df

random.seed(523)
test_neg_df = negative_sampling(test_df, set(data.dst_ind))
test_neg_user_df = negative_sampling_by_user(test_df, set(data.dst_ind))

len(test_df), len(test_neg_df), len(test_neg_user_df)

def intersect(df1: pd.DataFrame, df2):
    df1 = df1.set_index(['src_ind', 'dst_ind'])
    df2 = df2.set_index(['src_ind', 'dst_ind'])
    return df1.join(df2, how='inner', lsuffix='_pos', rsuffix='_neg')

intersect(test_df, test_neg_df)

intersect(test_df, test_neg_user_df)

train_df.to_csv(path + 'train.csv', index=False)
valid_df.to_csv(path + 'valid.csv', index=False)
test_df.to_csv(path + 'test.csv', index=False)
test_neg_df.to_csv(path + 'test_neg.csv', index=False)
test_neg_user_df.to_csv(path + 'test_neg_user.csv', index=False)
