==> notebook/03-raw data analysis.ipynb <==
path = '../tyc_cm/'

# %load_ext autoreload
# %autoreload 2
import dgl
import torch

### brand
### brand-organize: 融资历程
### brand-brand: 竞品信息

# >>> graph_data = {
# ...    ('drug', 'interacts', 'drug'): (th.tensor([0, 1]), th.tensor([1, 2])),
# ...    ('drug', 'interacts', 'gene'): (th.tensor([0, 1]), th.tensor([2, 3])),
# ...    ('drug', 'treats', 'disease'): (th.tensor([1]), th.tensor([2]))
# ... }

brands = set()
brand_belong = []
brand_rongzi = []
brand_jingpin = []
k = 1
for line in open(path + 'brand_all.json'):
    line = eval(line)
    if k:
        print(line)
        k = 0
    brand_id = line['id']
    brands.add(brand_id)
    brand_belong.append((brand_id, line['公司ID'], 'brand_belong'))
    brand_rongzi.extend([(brand_id, i['id'], 'brand_rongzi') for d in line.get('融资历程', []) for i in d['投资方']])
    brand_jingpin.extend([(brand_id, i, 'brand_jingpin') for i in line.get('竞品信息', [])])

brand_graph = [brand_belong, brand_rongzi, brand_jingpin]

len(brands), len(brand_belong), len(brand_rongzi), len(brand_jingpin)

brand_rongzi[0]

### company
### 股东信息: company-human, company-company
### 对外投资: company-company
### 竞品信息: company-company
### 供应商: company-company
### 历史股东: company-human, company-company

comps = set()
comp_gudong = []
comp_touzi = []
comp_jingpin = []
comp_gongying = []
comp_lsgudong = []
k = 1

for line in open('../data/raw/company_total.json'):
    line = eval(line)
    if k:
        print(line)
        print(line.keys())
        k = 0
    comp_id = line['id']
    comps.add(comp_id)
    comp_gudong.extend([(comp_id, d.get('CompanyID') or d.get('HumanID') if isinstance(d, dict) else d, 'comp_gudong') for d in line.get('股东信息', [])])
    comp_touzi.extend([(comp_id, d.get('CompanyID') or d.get('HumanID') if isinstance(d, dict) else d, 'comp_touzi') for d in line.get('对外投资', [])])
    comp_jingpin.extend([(comp_id, d.get('CompanyID') if isinstance(d, dict) else d, 'comp_jingpin') for d in line.get('竞品信息', [])])
    comp_gongying.extend([(comp_id, d.get('CompanyID') if isinstance(d, dict) else d, 'comp_gongying') for d in line.get('供应商', [])])
    comp_lsgudong.extend([(comp_id, d.get('CompanyID') or d.get('HumanID'), 'comp_lsgudong') for d in line.get('历史股东', [])])

comp_graph = [comp_gudong, comp_touzi, comp_jingpin, comp_gongying, comp_lsgudong]

len(comps), len(comp_gudong), len(comp_touzi), len(comp_jingpin), len(comp_gongying), len(comp_lsgudong)

len(set(list(zip(*brand_belong))[1]))

bs, cs, _ = list(zip(*brand_belong))

dd = {k: [] for k in cs}

for b, c, _ in brand_belong:
    dd[c].append(b)

dd

### organize
### 工商追踪: organize-company
### 公开投资事件: organize-company
### 未公开投资: organize-company

orgs = set()
org_gongshang = []
org_gktouzi = []
org_wgktouzi = []
k = 1

for line in open('../data/raw/organize_all.json'):
    line = eval(line)
    if k:
        print(line)
        print(line.keys())
        k = 0
    org_id = line['id']
    orgs.add(org_id)
    org_gongshang.extend([(org_id, d.get('CompanyID'), 'org_gongshang') for d in line.get('工商追踪', [])])
    org_gktouzi.extend([(org_id, d.get('CompanyID'), 'org_gktouzi') for d in line.get('公开投资事件', [])])
    org_wgktouzi.extend([(org_id, i.get('企业名称ID'), 'org_wgktouzi') for d in line.get('未公开投资', []) for i in d['投资方']])

org_graph = [org_gongshang, org_gktouzi, org_wgktouzi]

len(orgs), len(org_gongshang), len(org_gktouzi), len(org_wgktouzi)

### 对节点编号

brand_ord = {d: i for i, d in enumerate(brands)}
comp_ord = {d: i for i, d in enumerate(comps)}
org_ord = {d: i for i, d in enumerate(orgs)}

type_set = {'brand', 'company', 'organize'}
type_ord = {'brand': brand_ord, 'company': comp_ord, 'organize': org_ord}

graph_edge = {}

for graph in [brand_graph, comp_graph, org_graph]:
    for edges in graph:
        for u, v, et in edges:
            if v is None:
                continue
            ut, vt = u.split('/')[0], v.split('/')[0]
            # 类型不符合
            if ut not in type_set or vt not in type_set:
                continue
            # 公司/组织/品牌未爬取
            if u not in type_ord[ut] or v not in type_ord[vt]:
                continue
            if (ut, et, vt) in graph_edge:
                graph_edge[(ut, et, vt)].append((type_ord[ut][u], type_ord[vt][v]))
            else:
                graph_edge[(ut, et, vt)] = [(type_ord[ut][u], type_ord[vt][v])]

graph_edge.keys()

graph_data = {k: (torch.tensor(v)[:, 0], torch.tensor(v)[:, 1]) for k, v in graph_edge.items()}

g = dgl.heterograph(graph_data)

g.ntypes, g.etypes, g.canonical_etypes

g

torch.save([g, type_ord], path + 'heter_graph.pt')

g.num_nodes(), g.num_edges()

(51039+1147103+150568)/55056

1147103/55056, 51039/55056, 150568/55056

[(67486+70566+252377+15120+70047)/92524, 275658/92524]

[(135706+49446+103839)/14290]

f = open(path + 'brand_all.json')
brand_labels = []
brand_members = []
for line in f:
    line = eval(line)
    brand_labels.extend(line['标签'])
    if '核心成员' in line:
        brand_members.extend(line['核心成员'])

from collections import Counter

sorted(dict(Counter(brand_labels)).items(), key=lambda d: d[1], reverse=True)

len(dict(Counter(brand_labels)))

len(brand_members)

185944/55056

f = open(path + 'company_total.json')
eval(next(f))['企业业务']

"仅20年，互联网就完成了“传统IT行业—互联网—移动互联网—人工智能—IOT”的更迭，技术直接对世界带来影响所用的时间从10年缩至5年，且速度越来越快。优课达团队希望通过累积的经验，为热爱编程的同学们规划科学的学习路径，更快更好地学写代码；通过各种测试让同学们对自己有更加清晰的认知； 通过简历指导和大厂内推，让每个优秀的人都有闪光的机会！".count

f = open(path + 'company_total.json')
comp_label = []

for line in f:
    line = eval(line)
    if '企业业务' in line:
        comp_label.extend([d['标签'] for d in line['企业业务'] if '标签' in d])

len(Counter(comp_label))

sorted(dict(Counter(comp_label)).items(), key=lambda d: d[1], reverse=True)

comp_members = [i for i in list(zip(*comp_gudong))[1] if i is not None and 'human' in i]

len(comp_members)

len([i for i in list(zip(*comp_lsgudong))[1] if i is not None and 'human' in i])

(210000+93789)/92524

edges = {('brand', 'brand_belong', 'company'): 51039, ('brand', 'brand_jingpin', 'brand'): 1147103, ('brand', 'brand_rongzi', 'organize'): 150568, ('company', 'comp_gongying', 'company'): 67486, ('company', 'comp_gudong', 'company'): 70566, ('company', 'comp_jingpin', 'brand'): 275658, ('company', 'comp_jingpin', 'company'): 252377, ('company', 'comp_lsgudong', 'company'): 15120, ('company', 'comp_touzi', 'company'): 70047, ('organize', 'org_gktouzi', 'company'): 135706, ('organize', 'org_gongshang', 'company'): 49446, ('organize', 'org_wgktouzi', 'company'): 103839}

sorted(dict(Counter(edges)).items(), key=lambda d: d[1], reverse=True)

-
